Use fast map (64bits -> 64bits)

+++ internal/pathfinding/heuristics.go
@@ -6,21 +6,21 @@ import (
 )
 
 // Heuristic An heuristic function evaluates a given position and help us choose which path to explore first.
-type Heuristic func(p world.Position, goal world.Position) float32
+type Heuristic func(p world.Position, goal world.Position) float64
 
 // Manhattan Adapted when moving in 4 directions.
-func Manhattan(p world.Position, goal world.Position) float32 {
-	return float32(math.Abs(float64(p.X-goal.X)) + math.Abs(float64(p.Y-goal.Y)))
+func Manhattan(p world.Position, goal world.Position) float64 {
+	return math.Abs(float64(p.X-goal.X)) + math.Abs(float64(p.Y-goal.Y))
 }
 
 // Euclidean Adapted when moving in 8 directions.
-func Euclidean(p world.Position, goal world.Position) float32 {
-	return float32(math.Sqrt(float64(EuclideanSq(p, goal))))
+func Euclidean(p world.Position, goal world.Position) float64 {
+	return math.Sqrt(EuclideanSq(p, goal))
 }
 
 // EuclideanSq Same as Euclidean but save us a root square.
-func EuclideanSq(p world.Position, goal world.Position) float32 {
+func EuclideanSq(p world.Position, goal world.Position) float64 {
 	dx := math.Abs(float64(p.X - goal.X))
 	dy := math.Abs(float64(p.Y - goal.Y))
-	return float32(dx*dx + dy*dy)
+	return dx*dx + dy*dy
 }
